store price as integer instead of float

+++ src/models/shop/inventory_models.py
@@ -24,7 +24,7 @@ class Product(ModelWithImage, db.Model):
 
     id = db.Column(db.Integer, primary_key=True, nullable=False)
     name = db.Column(db.String(40), nullable=False)
-    price = db.Column(db.Float, nullable=False)
+    price = db.Column(db.Integer, nullable=False)
     description = db.Column(db.String(500), nullable=False)
     _image = db.Column(db.String(36), nullable=False)
     stock = db.Column(db.Integer, nullable=False)
